Differentiate between element and attribute identifiers
