+++ src/AttributeIdentifierInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace webignition\DomElementIdentifier;
+
+use webignition\DomElementLocator\ElementLocatorInterface;
+
+interface AttributeIdentifierInterface extends ElementIdentifierInterface
+{
+    public function getAttributeName(): ?string;
+    public function withAttributeName(string $attributeName): ElementIdentifierInterface;
+}
